mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.3",
    "openai>=1.99.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
